Load auth manager in ha and core-edge
