+++ community/kernel/src/main/java/org/neo4j/kernel/impl/factory/CommunityEditionModule.java
@@ -29,7 +29,6 @@ import org.neo4j.io.pagecache.IOLimiter;
 import org.neo4j.io.pagecache.PageCache;
 import org.neo4j.kernel.DatabaseAvailability;
 import org.neo4j.kernel.NeoStoreDataSource;
-import org.neo4j.kernel.api.security.AuthManager;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.impl.api.SchemaWriteGuard;
 import org.neo4j.kernel.impl.api.index.RemoveOrphanConstraintIndexesOnStartup;
@@ -208,40 +207,6 @@ public class CommunityEditionModule extends EditionModule
         }
 
         throw new IllegalArgumentException( "No lock manager found with the name '" + key + "'." );
-    }
-
-    public static AuthManager createAuthManager( Config config, LogService logging )
-    {
-        boolean authEnabled = config.get( GraphDatabaseSettings.auth_enabled );
-        if ( !authEnabled )
-        {
-            return AuthManager.NO_AUTH;
-        }
-
-        String key = config.get( GraphDatabaseSettings.auth_manager );
-        for ( AuthManager.Factory candidate : Service.load( AuthManager.Factory.class ) )
-        {
-            String candidateId = candidate.getKeys().iterator().next();
-            if ( candidateId.equals( key ) )
-            {
-                return candidate.newInstance( config, logging.getUserLogProvider() );
-            }
-            else if ( key.equals( "" ) )
-            {
-                logging.getInternalLog( CommunityFacadeFactory.class )
-                        .info( "No auth manager implementation specified, defaulting to '" + candidateId + "'" );
-                return candidate.newInstance( config, logging.getUserLogProvider() );
-            }
-        }
-
-        if ( key.equals( "" ) )
-        {
-            logging.getUserLog( CommunityFacadeFactory.class )
-                    .error( "No auth manager implementation specified and no default could be loaded." );
-            throw new IllegalArgumentException( "No auth manager found." );
-        }
-
-        throw new IllegalArgumentException( "No auth manager found with the name '" + key + "'." );
     }
 
     protected TransactionHeaderInformationFactory createHeaderInformationFactory()
