+++ community/kernel/src/main/java/org/neo4j/kernel/impl/factory/EditionModule.java
@@ -20,8 +20,11 @@
 package org.neo4j.kernel.impl.factory;
 
 import org.neo4j.graphdb.DependencyResolver;
+import org.neo4j.graphdb.factory.GraphDatabaseSettings;
+import org.neo4j.helpers.Service;
 import org.neo4j.io.pagecache.IOLimiter;
 import org.neo4j.kernel.NeoStoreDataSource;
+import org.neo4j.kernel.api.security.AuthManager;
 import org.neo4j.kernel.configuration.Config;
 import org.neo4j.kernel.impl.api.CommitProcessFactory;
 import org.neo4j.kernel.impl.api.SchemaWriteGuard;
@@ -32,6 +35,7 @@ import org.neo4j.kernel.impl.core.RelationshipTypeTokenHolder;
 import org.neo4j.kernel.impl.coreapi.CoreAPIAvailabilityGuard;
 import org.neo4j.kernel.impl.factory.GraphDatabaseFacadeFactory.Configuration;
 import org.neo4j.kernel.impl.locking.Locks;
+import org.neo4j.kernel.impl.logging.LogService;
 import org.neo4j.kernel.impl.store.id.IdGeneratorFactory;
 import org.neo4j.kernel.impl.transaction.TransactionHeaderInformationFactory;
 import org.neo4j.kernel.info.DiagnosticsManager;
@@ -87,5 +91,39 @@ public abstract class EditionModule
         sysInfo.set( UsageDataKeys.edition, databaseInfo.edition );
         sysInfo.set( UsageDataKeys.operationalMode, databaseInfo.operationalMode );
         config.augment( singletonMap( Configuration.editionName.name(), databaseInfo.edition.toString() ) );
+    }
+
+    public static AuthManager createAuthManager( Config config, LogService logging )
+    {
+        boolean authEnabled = config.get( GraphDatabaseSettings.auth_enabled );
+        if ( !authEnabled )
+        {
+            return AuthManager.NO_AUTH;
+        }
+
+        String key = config.get( GraphDatabaseSettings.auth_manager );
+        for ( AuthManager.Factory candidate : Service.load( AuthManager.Factory.class ) )
+        {
+            String candidateId = candidate.getKeys().iterator().next();
+            if ( candidateId.equals( key ) )
+            {
+                return candidate.newInstance( config, logging.getUserLogProvider() );
+            }
+            else if ( key.equals( "" ) )
+            {
+                logging.getInternalLog( CommunityFacadeFactory.class )
+                        .info( "No auth manager implementation specified, defaulting to '" + candidateId + "'" );
+                return candidate.newInstance( config, logging.getUserLogProvider() );
+            }
+        }
+
+        if ( key.equals( "" ) )
+        {
+            logging.getUserLog( CommunityFacadeFactory.class )
+                    .error( "No auth manager implementation specified and no default could be loaded." );
+            throw new IllegalArgumentException( "No auth manager found." );
+        }
+
+        throw new IllegalArgumentException( "No auth manager found with the name '" + key + "'." );
     }
 }
