+++ enterprise/core-edge/pom.xml
@@ -113,6 +113,12 @@
 
       <dependency>
         <groupId>org.neo4j</groupId>
+        <artifactId>neo4j-security-enterprise</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>org.neo4j</groupId>
         <artifactId>neo4j-logging</artifactId>
         <version>${project.version}</version>
         <type>test-jar</type>
