+++ enterprise/core-edge/src/main/java/org/neo4j/coreedge/server/edge/EnterpriseEdgeEditionModule.java
@@ -115,6 +115,9 @@ public class EnterpriseEdgeEditionModule extends EditionModule
 
         life.add( dependencies.satisfyDependency(
                 new DefaultKernelData( fileSystem, pageCache, storeDir, config, graphDatabaseFacade ) ) );
+
+        life.add( dependencies.satisfyDependency( createAuthManager( config, logging ) ) );
+
         headerInformationFactory = TransactionHeaderInformationFactory.DEFAULT;
 
         schemaWriteGuard = () -> {};
